+++ addon/components/render-active-tooltips.js
@@ -1,8 +1,8 @@
-import Component from '@ember/component';
-import { inject } from '@ember/service';
 /* eslint-disable */
 import { readOnly } from '@ember/object/computed';
 /* eslint-enable */
+import Component from '@ember/component';
+import { inject } from '@ember/service';
 import layout from '../templates/components/render-active-tooltips';
 
 export default Component.extend({